+++ src/utils/ethereumUtils.ts
@@ -1,11 +1,9 @@
 import { BigNumberish } from '@ethersproject/bignumber';
 import { Provider } from '@ethersproject/providers';
 import { serialize } from '@ethersproject/transactions';
-import { Wallet } from '@ethersproject/wallet';
 import { ETH_ADDRESS as ETH_ADDRESS_AGGREGATORS } from '@rainbow-me/swaps';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { captureException } from '@sentry/react-native';
-import { mnemonicToSeed } from 'bip39';
 // @ts-expect-error ts-migrate(7016) FIXME: Could not find a declaration file for module 'eth-... Remove this comment to see the full error message
 import { parse } from 'eth-url-parser';
 import {
@@ -13,11 +11,10 @@ import {
   isValidAddress,
   toChecksumAddress,
 } from 'ethereumjs-util';
-import { hdkey } from 'ethereumjs-wallet';
 import { Contract } from 'ethers';
 import lang from 'i18n-js';
 import { isEmpty, isString, replace } from 'lodash';
-import { InteractionManager, Linking, NativeModules } from 'react-native';
+import { InteractionManager, Linking } from 'react-native';
 import { ETHERSCAN_API_KEY } from 'react-native-dotenv';
 import { useSelector } from 'react-redux';
 import URL from 'url-parse';
@@ -51,14 +48,6 @@ import {
   isZero,
   subtract,
 } from '@/helpers/utilities';
-import WalletTypes from '@/helpers/walletTypes';
-import {
-  DEFAULT_HD_PATH,
-  identifyWalletType,
-  WalletLibraryType,
-  EthereumPrivateKey,
-  EthereumWalletSeed,
-} from '@/model/wallet';
 import { Navigation } from '@/navigation';
 import { parseAssetNative } from '@/parsers';
 import store from '@/redux/store';
@@ -83,8 +72,6 @@ import {
 import Routes from '@/navigation/routesNames';
 import logger from '@/utils/logger';
 import { IS_IOS } from '@/env';
-
-const { RNBip39 } = NativeModules;
 
 const getNetworkNativeAsset = (
   network: Network
@@ -552,58 +539,6 @@ const checkIfUrlIsAScam = async (url: string) => {
   }
 };
 
-const deriveAccountFromMnemonic = async (mnemonic: string, index = 0) => {
-  let seed;
-  if (ios) {
-    seed = await mnemonicToSeed(mnemonic);
-  } else {
-    const res = await RNBip39.mnemonicToSeed({ mnemonic, passphrase: null });
-    seed = new Buffer(res, 'base64');
-  }
-  const hdWallet = hdkey.fromMasterSeed(seed);
-  const root = hdWallet.derivePath(DEFAULT_HD_PATH);
-  const child = root.deriveChild(index);
-  const wallet = child.getWallet();
-  return {
-    address: toChecksumAddress(wallet.getAddress().toString('hex')),
-    isHDWallet: true,
-    root,
-    type: WalletTypes.mnemonic,
-    wallet,
-    walletType: WalletLibraryType.bip39,
-  };
-};
-
-const deriveAccountFromPrivateKey = (privateKey: EthereumPrivateKey) => {
-  const ethersWallet = new Wallet(addHexPrefix(privateKey));
-  return {
-    address: ethersWallet.address,
-    isHDWallet: false,
-    root: null,
-    type: WalletTypes.privateKey,
-    wallet: ethersWallet,
-    walletType: WalletLibraryType.ethers,
-  };
-};
-
-const deriveAccountFromWalletInput = (input: EthereumWalletSeed) => {
-  const type = identifyWalletType(input);
-  if (type === WalletTypes.privateKey) {
-    return deriveAccountFromPrivateKey(input);
-  } else if (type === WalletTypes.readOnly) {
-    const ethersWallet = { address: addHexPrefix(input), privateKey: null };
-    return {
-      address: addHexPrefix(input),
-      isHDWallet: false,
-      root: null,
-      type: WalletTypes.readOnly,
-      wallet: ethersWallet,
-      walletType: WalletLibraryType.ethers,
-    };
-  }
-  return deriveAccountFromMnemonic(input);
-};
-
 function getBlockExplorer(network: Network) {
   switch (network) {
     case Network.mainnet:
@@ -826,9 +761,6 @@ const getBasicSwapGasLimit = (chainId: number) => {
 export default {
   calculateL1FeeOptimism,
   checkIfUrlIsAScam,
-  deriveAccountFromMnemonic,
-  deriveAccountFromPrivateKey,
-  deriveAccountFromWalletInput,
   formatGenericAsset,
   getAssetFromAllAssets,
   getAccountAsset,
